Refactor: Split ffx package into modules
Reorganizes the codebase for better maintainability:

- ffx/__init__.py    - Public API and exports
- ffx/exceptions.py  - Custom exception classes
- ffx/utils.py       - Utility functions (long_to_bytes, bytes_to_long)
- ffx/integer.py     - FFXInteger class
- ffx/encrypter.py   - FFXEncrypter class

All existing functionality and tests continue to work unchanged.
Import statements remain the same (import ffx).
